+++ packages/search/test/content/index.test.ts
@@ -288,6 +288,7 @@ describe("Content Search Service", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },
@@ -406,6 +407,7 @@ describe("Content Search Service", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },
@@ -513,6 +515,7 @@ describe("Content Search Service", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },
@@ -605,6 +608,7 @@ describe("Content Search Service", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },
@@ -701,6 +705,68 @@ describe("searchContent function", () => {
         groupIds: "any(foo,bar)",
         orgId: "any(be55891b4)",
       });
+    });
+  });
+
+  describe("allows a client to specify paging information", () => {
+    it("with the Hub API", async () => {
+      const pageString = common.btoa(
+        JSON.stringify({
+          hub: { start: 1, size: 10 },
+          ago: { start: 3, size: 20 },
+        })
+      );
+
+      const options: IContentSearchOptions = {
+        page: pageString,
+      };
+
+      const request: IContentSearchRequest = { options };
+
+      // Mock
+      const hubRequestMock = spyOn(common, "hubApiRequest").and.returnValue(
+        Promise.resolve({
+          data: [],
+        })
+      );
+
+      await searchContent(request);
+
+      expect(hubRequestMock).toHaveBeenCalledTimes(1);
+      expect(hubRequestMock.calls.argsFor(0)[1].params.page).toEqual({
+        key: pageString,
+      });
+    });
+
+    it("with the Portal API", async () => {
+      const pageParams = {
+        start: 10,
+        num: 24,
+      };
+
+      const options: IContentSearchOptions = {
+        isPortal: true,
+        page: common.btoa(JSON.stringify(pageParams)),
+      };
+
+      const request: IContentSearchRequest = { options };
+
+      // Mock
+      const searchItemsMock = spyOn(portal, "searchItems").and.returnValue(
+        Promise.resolve({
+          results: [],
+        })
+      );
+
+      await searchContent(request);
+
+      expect(searchItemsMock).toHaveBeenCalledTimes(1);
+      expect(searchItemsMock.calls.argsFor(0)[0].params.start).toEqual(
+        pageParams.start
+      );
+      expect(searchItemsMock.calls.argsFor(0)[0].params.num).toEqual(
+        pageParams.num
+      );
     });
   });
 
@@ -892,6 +958,7 @@ describe("searchContent function", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },
@@ -981,6 +1048,7 @@ describe("searchContent function", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },
@@ -1086,6 +1154,7 @@ describe("searchContent function", () => {
             agg: undefined,
             catalog: undefined,
             fields: undefined,
+            page: undefined,
             q: undefined,
           },
         },