+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [9.1.2] - October 14th 2021
+
+### Other Changes
+
+* Bug Fixes
+   * **hub-search**: searchContent now respects user-provided paging information [`cd24c6fdf`](https://github.com/Esri/hub.js/commit/cd24c6fdf3377a2bad6c422fb341f52a0aaba07f)
+
 ## [9.1.1] - October 12th 2021
 
 ### Other Changes
@@ -29,6 +36,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - **hub-search**: added export to searchDatasets function [`3e9ef3bda`](https://github.com/Esri/hub.js/commit/3e9ef3bda35c60e4569c2646f623f866f3cadf54)
 
 ## Unreleased
+
+
 
 ## [9.0.0] - October 8th 2021
 
@@ -64,7 +73,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Chores
 
-  - **hub-common**: allow commits that start w/ &amp;amp;amp;amp;amp;amp;#39;pr:&amp;amp;amp;amp;amp;amp;#39; to address PR review [`e344283f0`](https://github.com/Esri/hub.js/commit/e344283f00cd329effd45aa4eb6b004e78664de9)
+  - **hub-common**: allow commits that start w/ &amp;amp;amp;amp;amp;amp;amp;#39;pr:&amp;amp;amp;amp;amp;amp;amp;#39; to address PR review [`e344283f0`](https://github.com/Esri/hub.js/commit/e344283f00cd329effd45aa4eb6b004e78664de9)
 
 - Bug Fixes
   - **hub-content**: getContentFromPortal() returns the layer content when passed a hubId [`157fa1110`](https://github.com/Esri/hub.js/commit/157fa1110546238d6a63b410ec129b93db15e6aa)
@@ -323,7 +332,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - **hub-teams**: add canEditTeam [`d2c9fecf`](https://github.com/Esri/hub.js/commit/d2c9fecf0314e35ee5c1acb9583684f18740f0b1)
   - **hub-teams**: add username check [`36a8b792`](https://github.com/Esri/hub.js/commit/36a8b7924ef7cf11ac2f2e81abe0d9ca49e86ec7)
   - **hub-teams**: add canEditTeam tests [`d4d52c18`](https://github.com/Esri/hub.js/commit/d4d52c1879c01f44770c16de242822ce41ba2708)
-  - **hub-teams**: add case when user is not auth&amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;#39;d [`295e04a7`](https://github.com/Esri/hub.js/commit/295e04a7189151c827a25dcf9222c76448df71f7)
+  - **hub-teams**: add case when user is not auth&amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;#39;d [`295e04a7`](https://github.com/Esri/hub.js/commit/295e04a7189151c827a25dcf9222c76448df71f7)
 
 ## [8.7.0] - June 25th 2021
 
@@ -1143,7 +1152,11 @@ Attempt to address failed release
 
 ## [4.1.0] - May 21st 2020
 
+
+
 ## [4.0.0] - May 20th 2020
+
+
 
 ## [3.9.3] - May 14th 2020
 
@@ -1181,6 +1194,8 @@ Attempt to address failed release
   - **peerDependencies**: fix invalid peerDependencies [`08fae37`](https://github.com/Esri/hub.js/commit/08fae37875aa3e68a82970a54eedcbe13ef001a8) [#251](https://github.com/Esri/hub.js/issues/251)
 
 ## [3.9.1] - May 6th 2020
+
+
 
 ## [3.9.0] - May 6th 2020
 
@@ -1222,6 +1237,8 @@ Attempt to address failed release
 
 ## [3.7.4] - April 30th 2020
 
+
+
 ## [3.7.3] - April 29th 2020
 
 - New Features
@@ -1232,6 +1249,8 @@ Attempt to address failed release
 ### @esri/hub-common
 
 ## [3.7.1] - April 28th 2020
+
+
 
 ## [3.7.0] - April 28th 2020
 
@@ -1287,7 +1306,11 @@ Attempt to address failed release
 
 ## [3.6.3] - January 27th 2020
 
+
+
 ## [3.6.2] - January 2nd 2020
+
+
 
 ## [3.6.0] - November 18th 2019
 
@@ -1381,6 +1404,8 @@ Attempt to address failed release
   - **search**: support filters specified implicitly
 
 ## [2.6.0] - August 20th 2019
+
+
 
 ## [2.5.0] - August 15th 2019
 
@@ -1994,4 +2019,5 @@ Initial Release
 [9.0.1]: https://github.com/Esri/hub.js/compare/v9.0.0...v9.0.1 "v9.0.1"
 [9.1.0]: https://github.com/Esri/hub.js/compare/v9.0.1...v9.1.0 "v9.1.0"
 [9.1.1]: https://github.com/Esri/hub.js/compare/v9.1.0...v9.1.1 "v9.1.1"
-[HEAD]: https://github.com/Esri/hub.js/compare/v9.1.1...HEAD "Unreleased Changes"
+[9.1.2]: https://github.com/Esri/hub.js/compare/v9.1.1...v9.1.2 "v9.1.2"
+[HEAD]: https://github.com/Esri/hub.js/compare/v9.1.2...HEAD "Unreleased Changes"